Envois par mail et selection adresses

+++ noethys/Dlg/DLG_Envoi_sms.py
@@ -262,8 +262,8 @@ def AjouteTexteImage(image=None, texte="", alignement="droite-bas", padding=0, t
         dc.DrawRoundedRectangleRect(wx.Rect(xRond, yRond, largeurRond, hauteurRond), hauteurRond / 2.0)
 
     # Texte
-    xTexte = xRond + largeurRond / 2.0 - largeurTexte / 2.0
-    yTexte = yRond + hauteurRond / 2.0 - hauteurTexte / 2.0 - 1
+    xTexte = int(xRond + largeurRond / 2.0 - largeurTexte / 2.0)
+    yTexte = int(yRond + hauteurRond / 2.0 - hauteurTexte / 2.0 - 1)
     dc.DrawText(texte, xTexte, yTexte)
 
     mdc.SelectObject(wx.NullBitmap)
